[orm] change properties namespaces
