+++ library/umi/orm/collection/BaseCollection.php
@@ -27,7 +27,7 @@ use umi\orm\metadata\field\relation\ManyToManyRelationField;
 use umi\orm\metadata\IMetadata;
 use umi\orm\metadata\IObjectType;
 use umi\orm\object\IObject;
-use umi\orm\object\property\ICalculableProperty;
+use umi\orm\object\property\calculable\ICalculableProperty;
 use umi\orm\object\property\IProperty;
 use umi\orm\persister\IObjectPersisterAware;
 use umi\orm\persister\TObjectPersisterAware;
