+++ library/umi/orm/object/property/calculable/CounterProperty.php
@@ -7,15 +7,15 @@
  * @license   http://umi-framework.ru/license/bsd-3 BSD-3 License
  */
 
-namespace umi\orm\object\property;
+namespace umi\orm\object\property\calculable;
 
 use umi\orm\metadata\field\IField;
 use umi\orm\object\IObject;
 
 /**
- * Класс свойства с вычисляемым значением
+ * Класс свойства счетчика.
  */
-class CalculableProperty extends BaseProperty implements ICalculableProperty
+class CounterProperty extends CalculableProperty implements ICounterProperty
 {
 
     /**
@@ -32,10 +32,21 @@ class CalculableProperty extends BaseProperty implements ICalculableProperty
     /**
      * {@inheritdoc}
      */
-    public function recalculate()
+    public function increment()
     {
-        $this->isModified = true;
-        $this->object->setIsModified();
+        $value = $this->getValue() + 1;
+        $this->setValue($value);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function decrement()
+    {
+        $value = $this->getValue() - 1;
+        $this->setValue($value);
 
         return $this;
     }
