+++ library/umi/orm/object/property/calculable/ICalculableProperty.php
@@ -7,7 +7,9 @@
  * @license   http://umi-framework.ru/license/bsd-3 BSD-3 License
  */
 
-namespace umi\orm\object\property;
+namespace umi\orm\object\property\calculable;
+
+use umi\orm\object\property\IProperty;
 
 /**
  * Иньерфейс для свойства с вычисляемым значением
