+++ library/umi/orm/object/property/localized/ILocalizedProperty.php
@@ -7,7 +7,9 @@
  * @license   http://umi-framework.ru/license/bsd-3 BSD-3 License
  */
 
-namespace umi\orm\object\property;
+namespace umi\orm\object\property\localized;
+
+use umi\orm\object\property\IProperty;
 
 /**
  * Интерфейс свойства, имеющего локализацию.
