+++ tests/utest/cache/unit/engine/MultiDbTest.php
@@ -9,7 +9,6 @@
 
 namespace utest\cache\unit\engine;
 
-use Doctrine\DBAL\Types\Type;
 use umi\cache\engine\Db;
 use umi\dbal\cluster\server\IServer;
