+++ tests/utest/orm/unit/object/property/LocalizedPropertyTest.php
@@ -11,7 +11,7 @@ namespace utest\orm\unit\object\property;
 
 use umi\orm\metadata\field\string\TextField;
 use umi\orm\object\IObject;
-use umi\orm\object\property\LocalizedProperty;
+use umi\orm\object\property\localized\LocalizedProperty;
 use utest\orm\ORMTestCase;
 
 /**
